pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Allen Anthes <allen@operationcode.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
bcrypt = "^4.0"
boto3 = "^1.34"
cryptography = "^41.0"
django = "^3.2"
djangorestframework = "^3.14"
djangorestframework-camel-case = "^1.3"
django-allauth = "^0.54"
django-anymail = "^10.1"
django-background-tasks = "^1.2"
django-cors-headers = "^4.0"
django-recaptcha2 = "^1.4"
dj-rest-auth = "^4.0"
django-split-settings = "^1.2"
django-storages = "^1.14"
django-widget-tweaks = "^1.4"
drf-yasg = "^1.21"
gunicorn = "^21.2"
packaging = "^23.0"
pillow = "^10.0"
psycopg2-binary = "^2.9"
python-decouple = "^3.8"
mailchimp3 = "^3.0"
mandrill = "^1.0"
python-dotenv = "^1.0"
requests = "^2.31"
sentry-sdk = "^1.38"
six = "^1.16"
django-allow-cidr = "^0.3.1"
django-health-check = "^3.16"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7"
black = "^23.0"
coverage = "^7.3"
django-debug-toolbar = "^4.2"
factory_boy = "^3.3"
flake8 = "^6.0"
flake8-bugbear = "^23.0"
isort = "^5.12"
pyhumps = "^3.0"
pytest = "^7.4"
pytest-django = "^4.5"
pytest-env = "^1.1"
pytest-mock = "^3.11"
pytest-pythonpath = "^0.7"
responses = "^0.23"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
